pyproject.toml
[tool.poetry]
name = "projectpoetry"
version = "0.1.0"
description = ""
authors = ["Agnehska <45421349+Agnehska@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.2"
black = "^24.4.2"
flake8 = "^7.0.0"
django = "^5.0.6"
pytest = "^8.2.1"
pytest-django = "^4.8.0"
pytest-mock = "^3.14.0"
djangorestframework = "^3.15.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.24.0"

[tool.black]
line-length = 88
exclude = '.venv'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
